+++ myems-api/reports/photovoltaicpowerstationdashboard.py
@@ -80,7 +80,7 @@ class Reporting:
         print(space_dict)
         # Get energy storage power stations
         query = (" SELECT m.id, m.name, m.uuid, "
-                 "        m.address, m.postal_code, m.latitude, m.longitude, "
+                 "        m.address, m.latitude, m.longitude, "
                  "        m.rated_capacity, m.rated_power, m.description, m.phase_of_lifecycle "
                  " FROM tbl_photovoltaic_power_stations m, tbl_photovoltaic_power_stations_users mu "
                  " WHERE m.id = mu.photovoltaic_power_station_id AND mu.user_id = %s "
@@ -110,13 +110,12 @@ class Reporting:
                                "uuid": row[2],
                                "address": row[3],
                                "space_name": space_dict.get(row[0]),
-                               "postal_code": row[4],
-                               "latitude": row[5],
-                               "longitude": row[6],
-                               "rated_capacity": row[7],
-                               "rated_power": row[8],
-                               "description": row[9],
-                               "phase_of_lifecycle": row[10],
+                               "latitude": row[4],
+                               "longitude": row[5],
+                               "rated_capacity": row[6],
+                               "rated_power": row[7],
+                               "description": row[8],
+                               "phase_of_lifecycle": row[9],
                                "status": 'online' if is_online else 'offline'}
                 total_rated_capacity += row[7]
                 total_rated_power += row[8]